Moved back relay entry to server package
